+++ src/main/java/enterprises/orbital/impl/evexmlapi/crp/ApiContainerLog.java
@@ -7,15 +7,15 @@ import enterprises.orbital.evexmlapi.crp.IContainerLog;
 public class ApiContainerLog implements IContainerLog {
   private Date   logTime;
   private long   itemID;
-  private long   itemTypeID;
+  private int    itemTypeID;
   private long   actorID;
   private String actorName;
   private byte   flag;
-  private int    locationID;
+  private long   locationID;
   private String action;
   private String passwordType;
-  private long   typeID;
-  private int    quantity;
+  private int    typeID;
+  private long   quantity;
   private String oldConfiguration;
   private String newConfiguration;
 
@@ -24,7 +24,8 @@ public class ApiContainerLog implements IContainerLog {
     return logTime;
   }
 
-  public void setLogTime(Date logTime) {
+  public void setLogTime(
+                         Date logTime) {
     this.logTime = logTime;
   }
 
@@ -33,16 +34,18 @@ public class ApiContainerLog implements IContainerLog {
     return itemID;
   }
 
-  public void setItemID(long itemID) {
+  public void setItemID(
+                        long itemID) {
     this.itemID = itemID;
   }
 
   @Override
-  public long getItemTypeID() {
+  public int getItemTypeID() {
     return itemTypeID;
   }
 
-  public void setItemTypeID(long itemTypeID) {
+  public void setItemTypeID(
+                            int itemTypeID) {
     this.itemTypeID = itemTypeID;
   }
 
@@ -51,7 +54,8 @@ public class ApiContainerLog implements IContainerLog {
     return actorID;
   }
 
-  public void setActorID(long actorID) {
+  public void setActorID(
+                         long actorID) {
     this.actorID = actorID;
   }
 
@@ -60,7 +64,8 @@ public class ApiContainerLog implements IContainerLog {
     return actorName;
   }
 
-  public void setActorName(String actorName) {
+  public void setActorName(
+                           String actorName) {
     this.actorName = actorName;
   }
 
@@ -69,16 +74,18 @@ public class ApiContainerLog implements IContainerLog {
     return flag;
   }
 
-  public void setFlag(byte flag) {
+  public void setFlag(
+                      byte flag) {
     this.flag = flag;
   }
 
   @Override
-  public int getLocationID() {
+  public long getLocationID() {
     return locationID;
   }
 
-  public void setLocationID(int locationID) {
+  public void setLocationID(
+                            long locationID) {
     this.locationID = locationID;
   }
 
@@ -87,7 +94,8 @@ public class ApiContainerLog implements IContainerLog {
     return action;
   }
 
-  public void setAction(String action) {
+  public void setAction(
+                        String action) {
     this.action = action;
   }
 
@@ -96,25 +104,28 @@ public class ApiContainerLog implements IContainerLog {
     return passwordType;
   }
 
-  public void setPasswordType(String passwordType) {
+  public void setPasswordType(
+                              String passwordType) {
     this.passwordType = passwordType;
   }
 
   @Override
-  public long getTypeID() {
+  public int getTypeID() {
     return typeID;
   }
 
-  public void setTypeID(long typeID) {
+  public void setTypeID(
+                        int typeID) {
     this.typeID = typeID;
   }
 
   @Override
-  public int getQuantity() {
+  public long getQuantity() {
     return quantity;
   }
 
-  public void setQuantity(int quantity) {
+  public void setQuantity(
+                          long quantity) {
     this.quantity = quantity;
   }
 
@@ -123,7 +134,8 @@ public class ApiContainerLog implements IContainerLog {
     return oldConfiguration;
   }
 
-  public void setOldConfiguration(String oldConfiguration) {
+  public void setOldConfiguration(
+                                  String oldConfiguration) {
     this.oldConfiguration = oldConfiguration;
   }
 
@@ -132,7 +144,8 @@ public class ApiContainerLog implements IContainerLog {
     return newConfiguration;
   }
 
-  public void setNewConfiguration(String newConfiguration) {
+  public void setNewConfiguration(
+                                  String newConfiguration) {
     this.newConfiguration = newConfiguration;
   }
 }
